requirements.txt
asgiref==3.7.2
Django==4.2.7
django-appconf==1.0.6
django-client-side-image-cropping==0.1.10
django-solo==2.1.0
django-tinymce==3.6.1
sqlparse==0.4.4
tzdata==2023.3
